refactor: null과 blank 에러를 공통 에러로 추출
