+++ src/main/java/cart/exception/common/CommonException.java
@@ -1,0 +1,8 @@
+package cart.exception.common;
+
+public class CommonException extends RuntimeException{
+
+    public CommonException(final String message) {
+        super(message);
+    }
+}
