+++ src/test/java/cart/entity/ProductTest.java
@@ -2,7 +2,7 @@ package cart.entity;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import cart.exception.product.NullOrBlankException;
+import cart.exception.common.NullOrBlankException;
 import cart.exception.product.PriceNotUnderZeroException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
